MS5 - backend/api + frontend files with api requests - Add public and private separated API endpoints
